Added delete, update and forms for update

+++ FurnitureShop/FurnitureShop/Models/Product.cs
@@ -5,8 +5,8 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public double Weight { get; set; }
+        public decimal Weight { get; set; }
         public string Barcode { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
     }
 }
